Changed approach to only step functions, because wrapper approach is not supported with current pipeline syntax. Refactored code to DockerCompose class in package eu.eosc-synergy.
